requirements.txt
amqp==2.3.2
Babel==2.6.0
billiard==3.5.0.4
braintree==3.49.0
cairocffi==0.9.0
CairoSVG==2.2.1
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
dj-database-url==0.5.0
django==2.1.3
django-contrib-comments==1.9.0
django-crispy-forms==1.7.2
django-localflavor==2.1
django-mathfilters==0.4.0
django-parler==1.9.2
django-rosetta==0.9.0
djangorestframework==3.9.0
flower==0.9.2
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
igdb-api-python==1.0
kombu==4.2.1
oauthlib==2.1.0
Pillow==5.3.0
polib==1.1.0
pycparser==2.19
PyJWT==1.6.4
Pyphen==0.9.5
PySocks==1.6.8
python-twitch-client==0.6.0
python3-openid==3.1.0
pytz==2018.7
redis==2.10.6
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
tinycss2==0.6.1
tornado==5.1.1
twilio==6.19.2
urllib3==1.24.1
vine==1.1.4
WeasyPrint==43
webencodings==0.5.1
whitenoise==4.1.1
psycopg2==2.7.5
